Splitter ut controller for filtervask i egen hjelpefil
